reorganize codegens
